лаб5/laba5/requirements.txt
beautifulsoup4==4.14.2
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comtypes==1.4.13
contourpy==1.3.3
cycler==0.12.1
Faker==38.2.0
Flask==3.1.2
fonttools==4.60.1
greenlet==3.2.4
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pluggy==1.6.0
pyfiglet==1.0.4
Pygments==2.19.2
pyparsing==3.2.5
pypiwin32==223
pytest==9.0.1
python-dateutil==2.9.0.post0
pyttsx3==2.99
pytz==2025.2
pywin32==311
PyYAML==6.0.3
random2==1.0.2
requests==2.32.5
six==1.17.0
soupsieve==2.8
SQLAlchemy==2.0.44
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
